[22] feat: add chart and fill it with dumies
